requirements.txt
amqp==5.1.1
aniso8601==9.0.1
asgiref==3.5.2
Babel==2.11.0
billiard==3.6.4.0
bleach==5.0.1
celery==5.2.7
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
dj-email-url==1.0.6
Django==4.1.3
django-appconf==1.0.5
django-cache-url==3.4.2
django-compressor==4.1
django-cors-headers==3.13.0
django-countries==7.4.2
django-filter==22.1
django-model-utils==4.2.0
django-phonenumber-field==7.0.0
django-prices==2.3.0
django-prices-openexchangerates==1.2.0
django-prices-vatlayer==1.1.0
django-simple-svg==0.2.1
django-storages==1.13.1
django-versatileimagefield==2.2
draftjs-sanitizer==1.0.0
enmerkar==0.7.1
graphene==3.1.1
graphene-django==3.0.0
graphql-core==3.2.3
graphql-relay==3.2.0
idna==3.4
Jinja2==3.1.2
jsonfield==3.1.0
kombu==5.2.4
MarkupSafe==2.1.1
oauthlib==3.2.2
phonenumbers==8.13.0
Pillow==9.3.0
prices==1.1.1
promise==2.3
prompt-toolkit==3.0.32
psycopg2==2.9.5
python-magic==0.4.27
python-magic-bin==0.4.14
pytz==2022.6
rcssmin==1.1.0
requests==2.28.1
requests-oauthlib==1.3.1
rjsmin==1.2.0
six==1.16.0
sqlparse==0.4.3
text-unidecode==1.3
typing_extensions==4.4.0
tzdata==2022.6
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
